requirements.txt
openai
beautifulsoup4
requests
python-dotenv
